Moved dataproviders from test class and minor fixes
